chore(server): move graphql subscription pubsub for kafka to server sharing the same Kafka global configuration
