+++ libs/util/nestjs/kafka/src/index.ts
@@ -3,4 +3,3 @@ export * from "./Kafka.module";
 export * from "./producer/";
 export * from "./pacemaker/";
 export * from "./kafka.transport";
-export * from "./kafka.pubsub.service";
