+++ packages/amplication-server/.env
@@ -3,8 +3,8 @@
 DEBUG_MODE=1
 NODE_ENV=development
 COMPOSE_PROJECT_NAME=amplication-server
-CUSTOM_ACTION_MIGRATION_TOKEN = 'CHANGE ME'
-FEATURE_ANNOUNCEMENT_NOTIFICATION_TOKEN = 'CHANGE ME'
+CUSTOM_ACTION_MIGRATION_TOKEN='CHANGE ME'
+FEATURE_ANNOUNCEMENT_NOTIFICATION_TOKEN='CHANGE ME'
 
 # Postgres
 POSTGRESQL_URL=postgresql://admin:admin@localhost:5432/amplication
@@ -77,6 +77,7 @@ CRON_SECRET_KEY=[cron-secret-key]
 
 FEATURE_CUSTOM_ACTIONS_ENABLED=true
 
-CHAT_OPENAI_KEY = "CHAT_OPENAI_KEY"
-CHAT_ASSISTANT_ID = "CHAT_ASSISTANT_ID"
-FEATURE_AI_ASSISTANT_ENABLED = false
+CHAT_OPENAI_KEY="CHAT_OPENAI_KEY"
+CHAT_ASSISTANT_ID="CHAT_ASSISTANT_ID"
+FEATURE_AI_ASSISTANT_ENABLED=true
+
